Changed types

+++ offline/add.cpp
@@ -2,27 +2,27 @@
 
 #include "../source.h"
 
-using namespace std;
+  using namespace std;
 
-Source add(vector<Source> sources){
-  // Initialize variables
-  num_sources  = sources.size();
-  num_channels = sources[0].num_channels;
-  num_samples  = sources[0].channels[0].size();
+  Source add(vector<Source> sources){
+    // Initialize variables
+    num_sources  = sources.size();
+    num_channels = sources[0].num_channels;
+    num_samples  = sources[0].channels[0].size();
 
-  // Reserve memory for result vec
-  Source result;
-  result.reserve(num_channels);
+    // Reserve memory for result vec
+    Source result;
+    result.reserve(num_channels);
 
-  for (int channel = 0; channel < num_channels; ++channel){
-    // New vector for new channel
-    vector<double> new_channel;
-    // Reserve memory for result vec<vec>
-    new_channel.reserve(num_sources * num_samples)
+    for (int channel = 0; channel < num_channels; ++channel){
+      // New vector for new channel
+      vector<float> new_channel;
+      // Reserve memory for result vec<vec>
+      new_channel.reserve(num_sources * num_samples)
 
-    // Concat sources' channels in result vector
-    for (int source = 0; source < num_sources; ++source){
-      vector<double> source_channel = multi_channels[source][channel];
+      // Concat sources' channels in result vector
+      for (int source = 0; source < num_sources; ++source){
+        vector<float> source_channel = multi_channels[source][channel];
       new_channel.insert(end(new_channel), begin(source_channel), end(source_channel));
     }
